# web/openapi/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Releaser.Server | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/admin/v1/applications/{appId}/audiences": {
      "get": {
        "tags": [
          "Audiences"
        ],
        "operationId": "ListAudiences",
        "parameters": [
          {
            "name": "appId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AudienceResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Audiences"
        ],
        "operationId": "CreateAudience",
        "parameters": [
          {
            "name": "appId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveAudienceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudienceResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/v1/applications/{appId}/audiences/{audienceId}": {
      "put": {
        "tags": [
          "Audiences"
        ],
        "operationId": "UpdateAudience",
        "parameters": [
          {
            "name": "appId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "audienceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveAudienceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudienceResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/HttpValidationProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Audiences"
        ],
        "operationId": "DeleteAudience",
        "parameters": [
          {
            "name": "appId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "audienceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "204": {
            "description": "No Content"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AudienceResponse": {
        "required": [
          "id",
          "name",
          "description",
          "includes",
          "excludes"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "includes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AudienceRule"
            }
          },
          "excludes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AudienceRule"
            }
          }
        }
      },
      "AudienceRule": {
        "required": [
          "kind"
        ],
        "type": "object",
        "anyOf": [
          {
            "$ref": "#/components/schemas/AudienceRuleEveryoneRule"
          },
          {
            "$ref": "#/components/schemas/AudienceRuleInstallationRule"
          },
          {
            "$ref": "#/components/schemas/AudienceRuleUserRule"
          },
          {
            "$ref": "#/components/schemas/AudienceRuleGroupRule"
          },
          {
            "$ref": "#/components/schemas/AudienceRuleCustomerRule"
          },
          {
            "$ref": "#/components/schemas/AudienceRuleAttributeRule"
          },
          {
            "$ref": "#/components/schemas/AudienceRulePlatformRule"
          },
          {
            "$ref": "#/components/schemas/AudienceRuleCurrentVersionRule"
          }
        ],
        "description": "One matching condition of an audience.",
        "discriminator": {
          "propertyName": "kind",
          "mapping": {
            "everyone": "#/components/schemas/AudienceRuleEveryoneRule",
            "installation": "#/components/schemas/AudienceRuleInstallationRule",
            "user": "#/components/schemas/AudienceRuleUserRule",
            "group": "#/components/schemas/AudienceRuleGroupRule",
            "customer": "#/components/schemas/AudienceRuleCustomerRule",
            "attribute": "#/components/schemas/AudienceRuleAttributeRule",
            "platform": "#/components/schemas/AudienceRulePlatformRule",
            "currentVersion": "#/components/schemas/AudienceRuleCurrentVersionRule"
          }
        }
      },
      "AudienceRuleAttributeRule": {
        "required": [
          "name",
          "values"
        ],
        "properties": {
          "kind": {
            "enum": [
              "attribute"
            ],
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "values": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "description": "Matches a verified custom attribute against a set of allowed values."
      },
      "AudienceRuleCurrentVersionRule": {
        "required": [
          "minimum",
          "maximumExclusive"
        ],
        "properties": {
          "kind": {
            "enum": [
              "currentVersion"
            ],
            "type": "string"
          },
          "minimum": {
            "type": [
              "null",
              "string"
            ]
          },
          "maximumExclusive": {
            "type": [
              "null",
              "string"
            ]
          }
        },
        "description": "Matches installations whose current version is within [Minimum, MaximumExclusive)."
      },
      "AudienceRuleCustomerRule": {
        "required": [
          "customerIds"
        ],
        "properties": {
          "kind": {
            "enum": [
              "customer"
            ],
            "type": "string"
          },
          "customerIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "AudienceRuleEveryoneRule": {
        "properties": {
          "kind": {
            "enum": [
              "everyone"
            ],
            "type": "string"
          }
        }
      },
      "AudienceRuleGroupRule": {
        "required": [
          "groups"
        ],
        "properties": {
          "kind": {
            "enum": [
              "group"
            ],
            "type": "string"
          },
          "groups": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "AudienceRuleInstallationRule": {
        "required": [
          "installationIds"
        ],
        "properties": {
          "kind": {
            "enum": [
              "installation"
            ],
            "type": "string"
          },
          "installationIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "AudienceRulePlatformRule": {
        "required": [
          "platforms"
        ],
        "properties": {
          "kind": {
            "enum": [
              "platform"
            ],
            "type": "string"
          },
          "platforms": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlatformTarget"
            }
          }
        }
      },
      "AudienceRuleUserRule": {
        "required": [
          "userIds"
        ],
        "properties": {
          "kind": {
            "enum": [
              "user"
            ],
            "type": "string"
          },
          "userIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "HttpValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      },
...
      "PlatformTarget": {
        "enum": [
          "Windows",
          "MacOS",
          "LinuxX64",
          "LinuxArm64",
          "LinuxArmv7l"
        ],
        "description": "Platform/architecture combination an updater asks for. Windows and macOS metadata cover all architectures in one file."
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "SaveAudienceRequest": {
        "required": [
          "name",
          "description",
          "includes",
          "excludes"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "includes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AudienceRule"
            }
          },
          "excludes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AudienceRule"
            }
          }
        }
      },
...
    }
  },
...
}